post component, reverting to old router setup, and cleaning code
