+++ web/src/Routes.tsx
@@ -1,14 +1,11 @@
 import './App.css';
-import { Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core';
-import {createBrowserHistory}  from 'history';
 import Home from './pages/home';
 import Register from './pages/register';
 import Login from './pages/login';
 import Profile from './pages/profile';
 import PrimaryAppBarContainer from './containers/primary-app-bar-container';
-
-const history = createBrowserHistory();
 
 const useStyles = makeStyles((theme) => ({
   content: {
@@ -21,7 +18,7 @@ const Routes = () => {
   const classes = useStyles();
 
   return (
-    <Router history={history}>
+    <BrowserRouter>
       <main className={classes.content}>
         <PrimaryAppBarContainer />
         <Switch>
@@ -31,7 +28,7 @@ const Routes = () => {
           <Route exact path="/login" component={Login} />
         </Switch>
       </main>
-    </Router>
+    </BrowserRouter>
   );
 }
