+++ web/src/containers/post-input-container/index.tsx
@@ -6,7 +6,9 @@ import { readGetUserPostsQuery } from '../../utils/cache-modification.util';
 import { currentGetUserPostsCursorVar, currentUserProfileVar, loggedInUserProfileVar } from '../../cache';
 
 const PostInputContainer = () => {
+  const loggedInUser: User = useReactiveVar(loggedInUserProfileVar);
   const currentUserProfile: User = useReactiveVar(currentUserProfileVar);
+  const currentGetUserPostsCursor: number = useReactiveVar(currentGetUserPostsCursorVar);
 
   const [createPost] = useMutation(CreatePostDocument, {
     update(cache, data) {
@@ -26,10 +28,10 @@ const PostInputContainer = () => {
   const handleCreatePost = async (e: React.FormEvent, callback: ()=> void): Promise<void> => {
     e.preventDefault();
     const formData = new FormData(e.target as HTMLFormElement);
-    currentGetUserPostsCursorVar(currentGetUserPostsCursorVar() + 1);
+    currentGetUserPostsCursorVar(currentGetUserPostsCursor + 1);
     await createPost({
       variables: {
-        creatorId: loggedInUserProfileVar().id,
+        creatorId: loggedInUser.id,
         content: formData.get('content') as string,
       }
     });
